1.x

+++ Runtime/Basics/AssetSystem.Get.cs
@@ -105,12 +105,12 @@ namespace AIO
 
 #if UNITY_IOS || UNITY_IPHONE
         [System.Runtime.InteropServices.DllImport("__Internal")]
-        private static extern double _IOS_GetFreeDiskSpace();
+        private static extern long _IOS_GetFreeDiskSpace();
 #endif
 
 #if UNITY_WEBGL
         // [System.Runtime.InteropServices.DllImport("__Internal")]
-        // private static extern double _WEBGL_GetFreeDiskSpace();
+        // private static extern long _WEBGL_GetFreeDiskSpace();
 #endif
 
         /// <summary>
@@ -136,7 +136,7 @@ namespace AIO
 #if UNITY_WEBGL
                 return 0;
 #elif UNITY_IOS || UNITY_IPHONE
-                return (long)_IOS_GetFreeDiskSpace();
+                return _IOS_GetFreeDiskSpace();
 #elif UNITY_ANDROID
                 var unityPluginLoader = new UnityEngine.AndroidJavaClass("IOHelper");
                 return unityPluginLoader.CallStatic<long>("GetFreeDiskSpace");
